Find missing and permute string
1)      Given a list of strings (single words or not), write a method to
return the longest, n-th string contained in the list. The method
signature is provided below:

private IList<string> getLongestString(int nTh, List<string> input){

};

If more than one string is of the same length and is the n-th longest,
return a collection.

2)      Given a list of numbers between some start and end value (i.e.
in a specific range), where all the numbers in that range are present
except for one, find the missing number

Ex:
[1,5,2,3]
Range is between 1 and 5
Missing number to  be returned is 4

private int findMissing(int[] input){

}